Add forcing to vpi_put_value for integers
This commit includes PR #6688.
It adds basic support for `vpiForceFlag` and `vpiReleaseFlag` to
`vpi_put_value` if `valuep->format valuep->format == vpiIntVal`. Any
other value formats are not supported and cause an error.
`vpi_get_value` is updated to return the value of either the signal
itself, or the `__VforceVal` control signal, depending on whether the
forcing is active.

The functionality is tested in the tests `t_vpi_force_minimal`, which
tests that forcing and releasing works on a clocked register being
forced with a VpiIntVal, as well as `t_vpi_forceable_bad`, which tests
that attempting to force a signal without marking it forceable causes an
error. The tests were run under Verilator (based on 048c97f with the
commit for isForceable applied), `Icarus Verilog version 13.0 (devel)
(s20251012-20-gcc496c3cf)`, and `xrun 24.09-a071`.
The goal is to develop the functionality in `t_vpi_force_minimal` into a
full-fledged `t_vpi_force` in the future, which tests the same data
types, value formats and scenarios as the existing
`t_vpi_put_value_array`.

The implementation is simply done using string concatenation of the
signal name with the __VforceEn and __VforceVal suffixes. While it works
and passes the newly implemented tests, this solution is quite brittle
and so I would prefer updating it to an approach that exposes
information about related force control signals during Verilation.

In order to accomplish this, I am considering extending
`VerilatedScope`'s `m_varsp` map to be not just a one-to-one mapping of
names to signals, but also provide meta-information about the signal,
such as the related force control signals in the case of a forceable
signal. I was wondering what the best way to go about this would be, and
I would appreciate your input here.

Additionally, `valuep` should get set to the value of the signal after
forcing when `vpi_put_value` is called with `vpiReleaseFlag`. This is
not yet implemented, since returning the proper value would require
knowing how the signal receives its value - through assignment or
clocking - and I have not been able to find this information easily.

Lastly, `vpi_get_value` simply re-implements a cumbersome variant of
what `__VforceRd` already does. It would be more elegant and reliable if
`vpi_get_value` could simply return this internal signal, but since it
cannot be marked `public_flat_rw` from the Verilog code, I have been
unable to access it.

I would greatly appreciate your input and pointers on how to move
forward with these issues.

Fixes #5933