Add vpi_put/vpi_get support for forcing signals
This commit includes PR #6688.
It adds support for `vpiForceFlag` and `vpiReleaseFlag` to
`vpi_put_value`. `vpi_get_value` is updated to return the value of
either the signal itself, or the `__VforceVal` control signal, depending
on whether the forcing is active.

The functionality is tested in the tests `t_vpi_force`, which tests that
forcing and releasing works on a clocked register being forced with a
VpiIntVal, as well as `t_vpi_forceable_bad`, which tests that attempting
to force a signal without marking it forceable causes an error. The
tests were run under Verilator (based on 91a59bbcc with the commit for
isForceable applied), `Icarus Verilog version 13.0 (devel)
(s20251012-20-gcc496c3cf)`, and `xrun 24.09-a071`. The goal is to
develop the functionality in `t_vpi_force` further in the future, in
order to test the same data types, value formats and scenarios as the
existing `t_vpi_put_value_array`.

The implementation is simply done using string concatenation of the
signal name with the __VforceEn and __VforceVal suffixes. The signal vop
that the vpi_put/get functions operate on is then pointed towards either
the base signal or the __VforceVal signal (for vpi_put) or the
__VforceRd signal (for vpi_get).

While it works and passes the newly implemented tests, this solution is
quite brittle and so an upcoming PR is going to present a proposal for
exposing information about related force control signals during
Verilation.

Additionally, `valuep` should get set to the value of the signal after
forcing when `vpi_put_value` is called with `vpiReleaseFlag`. This is
not yet implemented since `isContinuously` is not accessible in
Verilated code yet.

Lastly, `vpi_get_value` simply re-implements a cumbersome variant of
what `__VforceRd` already does. It would be more elegant and reliable if
`vpi_get_value` could simply return this internal signal, but since it
cannot be marked `public_flat_rw` from the Verilog code, this has been
postponed for a future PR.

Fixes #5933