Expand VerilatedVarProps to hold force info
This commit is an example of how changes to the verilated `__Syms` file
can be used to convey additional information about the relation of a
forceable signal to its force control signals without needing to rely on
string concatenation within `verilated_vpi.cpp`. This mockup shows how
PR #6704 could be simplified if a signal's `VerilatedVar` contained
information about its force control signals.

This is accomplished by adding two more member variables to
VerilatedVarProps, and one to VerilatedVar: A shared pointer to the
`__VforceRd` signal and a pair of raw pointers to the `__VforceEn` and
`__VforceVal` signals, as well as an `isContinuously` member. The
pointer to `__VforceRd` can be an owning pointer because the
`__VforceRd` signal cannot be set as `public_flat_rw` or
`public_flat_rd`, and so will never be added to the `VerilatedScope`.
Hence, the only instance of it exists in the `VerilatedVar` for its
forceable base signal. The pointers to the `__VforceEn` and
`__VforceVal` signals are already added to the `VerilatedScope` because
they are inherently public, so the base signal's `VerilatedVar` must
contain non-owning pointers to them.

These pointers are obtained by updating the `varInsert` member functions
to return a pointer to the newly created `VerilatedVar`, which is then
used as an argument in the new `forceableVarInsert`. In order to do
this, `V3EmitCSyms.cpp` was changed to ensure that the force control
signals are emitted _before_ the base signal.

Please note that this is an extremely hacky mockup. I committed horrific
crimes against code and strongly advise against merging this in its
current state. The point is to show how the implementation in
`verilated_vpi.cpp` can be simplified if _some_ way is found to retrieve
the force read and force control signals from the base signal's
`VerilatedVar`. The rest of the changes in the files other than
`verilated_vpi.cpp` are simply quick hacks to make this happen.

When considering what to change for a proper implementation, I am not
sure about the way to retrieve the `__VforceRd`, `__VforceEn` and
`__VforceVal` signals from the base signal's `VerilatedVpioVar`. I am
quite unhappy with adding the `m_forceReadSignalp` and
`m_forceControlSignals` member variables to the `VerilatedVarProps`,
because these variables take up space for every signal, forceable or
not, and the `forceReadSignal()` and `forceControlSignals()` functions
make no sense in the context of a non-forceable signal. My preferred
solution would have been to create a new `forceableVerilatedVar` which
inherits from `VerilatedVar` and contains these member variables and
functions. However, this would mean changing `VerilatedScope`'s
`m_varsp` to be a map between names and _pointers_ (preferably owning
`unique_ptr`s or `shared_ptr`s), which would of course mean modifying
all the accesses to this map, which was more refactoring than I was
comfortable with without checking in with you first.

The changes to `V3EmitCSyms.cpp` are simply a crude removal of the
sorting and changing the iteration over the signals to go backwards to
ensure that the insertion of the force read and force control signals
happen _before_ the base signal, so the `VerilatedVar`s are already
available for the `forceableVarInsert` function call. The insertion
itself happens by string concatenation again.

I am well aware that this simply kicks the string modification problem
down the road, and I have managed to make an implementation that is even
_more_ brittle than the previous one. A proper implementation should not
depend on the ordering of the `varInsert` calls nor string concatenation
within `V3EmitCSyms`, but rather should somehow embed the relation
between the base signal and the force control signals within the AST. I
have not been able to do that yet and would appreciate your input.