Fix vpi_get_value overwriting string outputs
When `vpi_get_value` is used to retrieve a string, i.e. `valuep->format`
is `vpiBinStrVal`, `vpiOctStrVal`, `vpiDecStrVal`, `vpiHexStrVal` or
`vpiStringVal`, it writes the result to the output parameter `valuep`'s
`value.str` as a `PLI_BYTE8*`.
The pointer that is written to this structure was previously a pointer
to the `static thread_local std::string t_outDynamicStr` variable. This
meant that the same pointer could be written to different `valuep`s, as
long as the `std::string`s data hasn't changed location, e.g. due to
resizing. The result of this is that a subsequent `vpi_get_value` could
change the string pointed to by an `s_vpi_value` obtained from an
earlier `vpi_get_value`.

IEEE 1800-2023, section 38.15, states that:
> The value shall be placed in an `s_vpi_value` structure, which has
been allocated by the application.

While this would be simple to implement by `strcpy`ing the result
directly into `valuep->value.str` and doing away with `t_outDynamicStr`
entirely, it is clear the current code does not respect this
requirement, and I suspect that existing user code may rely on the
current implementation and does not allocate memory for `value.str`.

In order to avoid breaking existing code, my proposed workaround is to
store the `t_outDynamicStr` in a static
`std::vector<std::unique_ptr<PLI_BYTE8[]>>` and then placing the
persistent address of the string stored inside that `unique_ptr` in the
output `s_vpi_value`. This ensures both that different `valuep`s point
to different `PLI_BYTE8*` addresses, as well as persistent lifetime.

To enable this, I also slightly changed `vl_strprintf`, which would
previously reset `t_outDynamicStr` to size 0, causing issues with the
new `storeOutputString` function which requires a correct `size`.

The downside is that I could not find a way to match the lifetime of the
items in the `unique_ptr`s to the lifetime of the `s_vpi_value`
structure, so this is essentially a memory leak because all the output
strings accumulate over the lifetime of the simulation with no way to
remove them. The usage of a `std::vector` of `unique_ptr`s instead of
`new std::string` allocations without `delete`s is to make the intent
clear and stop static code analysis warnings.

One alternate to reduce the memory footprint could be to only store
_unique_ output strings in a `std::set`. However, this would have the
potential to cause spooky action at a distance when a user re-uses a
`s_vpi_value` structure that has been used to retrieve a string with
`vpi_get_value`, because the user could directly manipulate the string
stored in the `static std::set`, which may also be referenced by other
`s_vpi_value`s that retrieved the same value.

I would appreciate your input on this.