Extend VerilatedVar to hold ForceableInfo pointer
This commit makes changes to the verilated `__Syms` file in order to
convey additional information about the relation of a forceable signal
to its force control signals without needing to rely on string
concatenation within `verilated_vpi.cpp`. This implementation builds
upon PR #6704 and simplifies the implementation within
`verilated_vpi.cpp`.

This is accomplished by adding a pointer to a `ForceableInfo` class
object to `VerilatedVar`. It contains pointers to the `VerilatedVar`s
for the `__VforceEn` and `__VforceVal` signals and owns a pointer to the
`__VforceRd` signal's `VerilatedVar`.

It obtains this information through an adapted `varInsert` function,
`forceableVarInsert`, emitted from `V3EmitCSyms.cpp`. There, if a
forceable signal is found, first the control signals are inserted into
the `m_varsp` map and their pointers are given to the the
`forceableVarInsert` function. Within that function, the `ForceableInfo`
object is constructed and ownership given to the base signal's
`VerilatedVar`.

The pointer to `__VforceRd` can be an owning pointer because the
`__VforceRd` signal cannot be set as `public_flat_rw` or
`public_flat_rd`, and so will never be added to the `VerilatedScope`.
Hence, the only instance of it exists in the `VerilatedVar` for its
forceable base signal. The pointers to the `__VforceEn` and
`__VforceVal` signals are already added to the `VerilatedScope` because
they are inherently public, so the base signal's `VerilatedVar`'s
`ForceableInfo` must contain non-owning pointers to them.

This required introducing an owning `unique_ptr` to a `ForceableInfo`
object as a member variable of `VerilatedVar`, which also necessitates
re-introducing `VerilatedVar`'s move constructor to allow it to be moved
into the `m_varsp` map.

Because `vpi_get_value` now retrieves the value of the `__VforceRd`
signal directly, rather than recreating it as in #6704, this also means
that a change made by `vpi_put_value` to a forceable signal is _not_
immediately visible to `vpi_get_value` anymore, but only after another
`eval`. This means that forceable and non-forceable variables behave
differently, and `t_vpi_var` would not pass anymore if the signals were
forceable, despite no forcing functionality being used. `t_vpi_force`
also had to be updated, requiring splitting up `putInertialDelay` to
check the outcome of the `vpi_put_value` in another function called
after an `eval` had happened.

In order to address this issue, I can think of three potential
approaches:
* Trigger an update of the `__VforceRd` signal from within
  `vpi_put_value`. However, seeing that `evalNeeded` exists, I suspect
  this is intentionally not done already for performance reasons.
* Keep the implementation from #6704 which re-created the `__VforceRd`
  signal, yielding instantaneous results. However, this would mean code
  duplication and possible issues if the computation of `__VforceRd`
  ever changes.
* Document that forceable and non-forceable signals behave differently.
  Yikes.

If you have any other ideas, I would greatly appreciate the input.